requirements.txt
python-telegram-bot
graphene
requests
pymongo
dnspython